# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdoptAgentRequest": {
        "properties": {
          "onConflict": {
            "anyOf": [
              {
                "enum": [
                  "keep_store",
                  "replace_store"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onconflict"
          }
        },
        "title": "AdoptAgentRequest",
        "type": "object"
      },
      "AdoptAgentResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "ref"
        ],
        "title": "AdoptAgentResponse",
        "type": "object"
      },
...
      "AgentActionsResponse": {
        "properties": {
          "canAdopt": {
            "title": "Canadopt",
            "type": "boolean"
          },
          "canDelete": {
            "title": "Candelete",
            "type": "boolean"
          }
        },
        "required": [
          "canAdopt",
          "canDelete"
        ],
        "title": "AgentActionsResponse",
        "type": "object"
      },
      "AgentBindingResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "state": {
            "enum": [
              "enabled",
              "disabled",
              "unsupported"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "state"
        ],
        "title": "AgentBindingResponse",
        "type": "object"
      },
      "AgentColumnResponse": {
        "properties": {
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "logoKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logokey"
          }
        },
        "required": [
          "harness",
          "label",
          "installed"
        ],
        "title": "AgentColumnResponse",
        "type": "object"
      },
      "AgentConfigEntryResponse": {
        "description": "One frontmatter key we do not interpret, shown verbatim.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "key",
          "value"
        ],
        "title": "AgentConfigEntryResponse",
        "type": "object"
      },
      "AgentDetailResponse": {
        "properties": {
          "canDelete": {
            "title": "Candelete",
            "type": "boolean"
          },
          "configuration": {
            "items": {
              "$ref": "#/components/schemas/AgentConfigEntryResponse"
            },
            "title": "Configuration",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "document": {
            "title": "Document",
            "type": "string"
          },
          "harnesses": {
            "items": {
              "$ref": "#/components/schemas/AgentHarnessDetailResponse"
            },
            "title": "Harnesses",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "storePath": {
            "title": "Storepath",
            "type": "string"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "ref",
          "name",
          "description",
          "prompt",
          "tools",
          "document",
          "storePath",
          "harnesses",
          "canDelete"
        ],
        "title": "AgentDetailResponse",
        "type": "object"
      },
      "AgentEntryResponse": {
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/AgentActionsResponse"
          },
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/AgentBindingResponse"
            },
            "title": "Bindings",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "harnessPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harnesspath"
          },
          "kind": {
            "enum": [
              "managed",
              "unmanaged"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "name",
          "description",
          "kind",
          "bindings",
          "actions"
        ],
        "title": "AgentEntryResponse",
        "type": "object"
      },
      "AgentHarnessDetailResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          },
          "installMethod": {
            "enum": [
              "symlink",
              "rendered",
              "none"
            ],
            "title": "Installmethod",
            "type": "string"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "logoKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logokey"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "state": {
            "enum": [
              "enabled",
              "disabled",
              "unsupported"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "label",
          "state",
          "path",
          "installMethod",
          "installed"
        ],
        "title": "AgentHarnessDetailResponse",
        "type": "object"
      },
      "AgentHarnessRequest": {
        "properties": {
          "harness": {
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness"
        ],
        "title": "AgentHarnessRequest",
        "type": "object"
      },
      "AgentInventoryResponse": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/AgentColumnResponse"
            },
            "title": "Columns",
            "type": "array"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AgentEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/AgentIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "columns",
          "entries"
        ],
        "title": "AgentInventoryResponse",
        "type": "object"
      },
      "AgentIssueResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "name",
          "reason"
        ],
        "title": "AgentIssueResponse",
        "type": "object"
      },
      "AgentMutationFailureResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "harness": {
            "title": "Harness",
            "type": "string"
          }
        },
        "required": [
          "harness",
          "error"
        ],
        "title": "AgentMutationFailureResponse",
        "type": "object"
      },
...
      "CreateAgentRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prompt": {
            "default": "",
            "title": "Prompt",
            "type": "string"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateAgentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "SetAgentHarnessesRequest": {
        "properties": {
          "harnesses": {
            "items": {
              "type": "string"
            },
            "title": "Harnesses",
            "type": "array"
          }
        },
        "title": "SetAgentHarnessesRequest",
        "type": "object"
      },
      "SetAgentHarnessesResultResponse": {
        "properties": {
          "failed": {
            "items": {
              "$ref": "#/components/schemas/AgentMutationFailureResponse"
            },
            "title": "Failed",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "succeeded": {
            "items": {
              "type": "string"
            },
            "title": "Succeeded",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "succeeded",
          "failed"
        ],
        "title": "SetAgentHarnessesResultResponse",
        "type": "object"
      },
...
      "UpdateAgentRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          }
        },
        "title": "UpdateAgentRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "harness-asset-manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents": {
      "get": {
        "operationId": "list_agents_api_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentInventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Agents",
        "tags": [
          "Agents"
        ]
      },
      "post": {
        "operationId": "create_agent_api_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent",
        "tags": [
          "Agents"
        ]
      }
    },
...
    "/api/agents/{agent_ref}": {
      "delete": {
        "operationId": "delete_agent_api_agents__agent_ref__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent",
        "tags": [
          "Agents"
        ]
      },
      "get": {
        "operationId": "get_agent_api_agents__agent_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent",
        "tags": [
          "Agents"
        ]
      },
      "put": {
        "operationId": "update_agent_api_agents__agent_ref__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent",
        "tags": [
          "Agents"
        ]
      }
    },
    "/api/agents/{agent_ref}/adopt": {
      "post": {
        "operationId": "adopt_agent_api_agents__agent_ref__adopt_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptAgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adopt Agent",
        "tags": [
          "Agents"
        ]
      }
    },
    "/api/agents/{agent_ref}/disable": {
      "post": {
        "operationId": "disable_agent_api_agents__agent_ref__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentHarnessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Agent",
        "tags": [
          "Agents"
        ]
      }
    },
    "/api/agents/{agent_ref}/enable": {
      "post": {
        "operationId": "enable_agent_api_agents__agent_ref__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentHarnessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Agent",
        "tags": [
          "Agents"
        ]
      }
    },
    "/api/agents/{agent_ref}/set-harnesses": {
      "post": {
        "operationId": "set_agent_harnesses_api_agents__agent_ref__set_harnesses_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_ref",
            "required": true,
            "schema": {
              "title": "Agent Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAgentHarnessesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetAgentHarnessesResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Agent Harnesses",
        "tags": [
          "Agents"
        ]
      }
    },
...
  }
}